refactor: extract StoreValidator from near-store
store-validator: all logic for validation,
depends on store and chain-configs
store-validator-bin: standalone binary,
depends on neard and store-validator-lib

Test plan
---------
N/A
