+++ chain/chain/Cargo.toml
@@ -29,6 +29,9 @@ near-pool = { path = "../pool" }
 [dev-dependencies]
 near-logger-utils = {path = "../../test-utils/logger"}
 
+[dev-dependencies]
+near-store-validator = { path = "../../test-utils/store-validator" }
+
 [features]
 # if enabled, we assert in most situations that are impossible unless some byzantine behavior is observed.
 byzantine_asserts = []
