+++ test-utils/store-validator/Cargo.toml
@@ -1,22 +1,10 @@
 [package]
-name = "store-validator"
+name = "near-store-validator"
 version = "0.1.0"
 authors = ["Near Inc <hello@nearprotocol.com>"]
 edition = "2018"
 
 [dependencies]
-ansi_term = "0.11"
-clap = "2.33"
-
-borsh = "0.5.0"
-
 near-chain-configs = { path = "../../core/chain-configs" }
-near-logger-utils = {path = "../../test-utils/logger"}
 near-primitives = { path = "../../core/primitives" }
 near-store = { path = "../../core/store" }
-near-chain = { path = "../../chain/chain" }
-neard = { path = "../../neard" }
-
-[dev-dependencies]
-serde_json = "1"
-near-client = { path = "../../chain/client" }
