+++ test-utils/store-validator/src/validate.rs
@@ -1,18 +1,16 @@
 use std::collections::{HashMap, HashSet};
 use std::convert::TryFrom;
-use std::sync::Arc;
 
 use borsh::BorshDeserialize;
 
-use near_chain_configs::GenesisConfig;
 use near_primitives::block::{Block, BlockHeader};
+use near_primitives::borsh;
 use near_primitives::hash::CryptoHash;
 use near_primitives::sharding::{ChunkHash, ShardChunk};
 use near_primitives::types::{BlockHeight, EpochId};
 use near_primitives::utils::index_to_bytes;
-
 #[allow(unused)]
-use crate::{
+use near_store::{
     read_with_cache, ColBlock, ColBlockExtra, ColBlockHeader, ColBlockHeight, ColBlockMisc,
     ColBlockPerHeight, ColBlockRefCount, ColBlocksToCatchup, ColChallengedBlocks, ColChunkExtra,
     ColChunkPerHeightShard, ColChunks, ColEpochLightClientBlocks, ColIncomingReceipts,
@@ -22,6 +20,8 @@ use crate::{
     KeyForStateChanges, ShardTries, Store, StoreUpdate, Trie, TrieChanges, TrieIterator,
     WrappedTrieChanges, TAIL_KEY,
 };
+
+use crate::StoreValidator;
 
 macro_rules! get_parent_function_name {
     () => {{
@@ -48,14 +48,17 @@ macro_rules! unwrap_or_err {
 }
 
 // All validations start here
-//
 
-fn nothing(_sv: &StoreValidator, _key: &[u8], _value: &[u8]) -> Result<(), String> {
+pub(crate) fn nothing(_sv: &StoreValidator, _key: &[u8], _value: &[u8]) -> Result<(), String> {
     // Make sure that validation is executed
     Ok(())
 }
 
-fn block_header_validity(_sv: &StoreValidator, key: &[u8], value: &[u8]) -> Result<(), String> {
+pub(crate) fn block_header_validity(
+    _sv: &StoreValidator,
+    key: &[u8],
+    value: &[u8],
+) -> Result<(), String> {
     let block_hash = CryptoHash::try_from(key.as_ref()).unwrap();
     match BlockHeader::try_from_slice(value) {
         Ok(header) => {
@@ -69,7 +72,11 @@ fn block_header_validity(_sv: &StoreValidator, key: &[u8], value: &[u8]) -> Resu
     }
 }
 
-fn block_hash_validity(_sv: &StoreValidator, key: &[u8], value: &[u8]) -> Result<(), String> {
+pub(crate) fn block_hash_validity(
+    _sv: &StoreValidator,
+    key: &[u8],
+    value: &[u8],
+) -> Result<(), String> {
     let block_hash = CryptoHash::try_from(key.as_ref()).unwrap();
     match Block::try_from_slice(value) {
         Ok(block) => {
@@ -83,7 +90,11 @@ fn block_hash_validity(_sv: &StoreValidator, key: &[u8], value: &[u8]) -> Result
     }
 }
 
-fn block_header_exists(sv: &StoreValidator, key: &[u8], _value: &[u8]) -> Result<(), String> {
+pub(crate) fn block_header_exists(
+    sv: &StoreValidator,
+    key: &[u8],
+    _value: &[u8],
+) -> Result<(), String> {
     let block_hash = CryptoHash::try_from(key.as_ref()).unwrap();
     match sv.store.get_ser::<BlockHeader>(ColBlockHeader, block_hash.as_ref()) {
         Ok(Some(_header)) => Ok(()),
@@ -92,7 +103,11 @@ fn block_header_exists(sv: &StoreValidator, key: &[u8], _value: &[u8]) -> Result
     }
 }
 
-fn chunk_hash_validity(_sv: &StoreValidator, key: &[u8], value: &[u8]) -> Result<(), String> {
+pub(crate) fn chunk_hash_validity(
+    _sv: &StoreValidator,
+    key: &[u8],
+    value: &[u8],
+) -> Result<(), String> {
     let chunk_hash = ChunkHash::try_from_slice(key.as_ref()).unwrap();
     match ShardChunk::try_from_slice(value) {
         Ok(shard_chunk) => {
@@ -106,7 +121,11 @@ fn chunk_hash_validity(_sv: &StoreValidator, key: &[u8], value: &[u8]) -> Result
     }
 }
 
-fn block_of_chunk_exists(sv: &StoreValidator, _key: &[u8], value: &[u8]) -> Result<(), String> {
+pub(crate) fn block_of_chunk_exists(
+    sv: &StoreValidator,
+    _key: &[u8],
+    value: &[u8],
+) -> Result<(), String> {
     let shard_chunk =
         unwrap_or_err!(ShardChunk::try_from_slice(value), "Can't deserialize ShardChunk");
     let height = shard_chunk.header.height_included;
@@ -138,7 +157,11 @@ fn block_of_chunk_exists(sv: &StoreValidator, _key: &[u8], value: &[u8]) -> Resu
     }
 }
 
-fn block_height_cmp_tail(sv: &StoreValidator, key: &[u8], value: &[u8]) -> Result<(), String> {
+pub(crate) fn block_height_cmp_tail(
+    sv: &StoreValidator,
+    key: &[u8],
+    value: &[u8],
+) -> Result<(), String> {
     let block_hash = CryptoHash::try_from(key.as_ref()).unwrap();
     let tail = match sv.store.get_ser::<BlockHeight>(ColBlockMisc, TAIL_KEY) {
         Ok(Some(tail)) => tail,
@@ -164,7 +187,7 @@ fn block_height_cmp_tail(sv: &StoreValidator, key: &[u8], value: &[u8]) -> Resul
     }
 }
 
-fn chunks_state_roots_in_trie(
+pub(crate) fn chunks_state_roots_in_trie(
     sv: &StoreValidator,
     _key: &[u8],
     value: &[u8],
@@ -179,91 +202,4 @@ fn chunks_state_roots_in_trie(
         unwrap_or_err!(item, format!("Can't find ShardChunk {} in Trie", shard_chunk));
     }
     Ok(())
-}
-
-//
-// All validations end here
-
-#[derive(Debug)]
-pub struct ErrorMessage {
-    pub col: DBCol,
-    pub msg: String,
-}
-
-impl ErrorMessage {
-    fn new(col: DBCol, msg: String) -> Self {
-        Self { col, msg }
-    }
-}
-
-pub struct StoreValidator {
-    config: GenesisConfig,
-    shard_tries: ShardTries,
-    store: Arc<Store>,
-
-    pub errors: Vec<ErrorMessage>,
-    tests: u64,
-}
-
-impl StoreValidator {
-    pub fn new(config: GenesisConfig, shard_tries: ShardTries, store: Arc<Store>) -> Self {
-        StoreValidator {
-            config,
-            shard_tries: shard_tries.clone(),
-            store: store.clone(),
-            errors: vec![],
-            tests: 0,
-        }
-    }
-    pub fn is_failed(&self) -> bool {
-        self.tests == 0 || self.errors.len() > 0
-    }
-    pub fn num_failed(&self) -> u64 {
-        self.errors.len() as u64
-    }
-    pub fn tests_done(&self) -> u64 {
-        self.tests
-    }
-    pub fn validate(&mut self) {
-        self.check(&nothing, &[0], &[0], ColBlockMisc);
-        for (key, value) in self.store.clone().iter(ColBlockHeader) {
-            // Block Header Hash is valid
-            self.check(&block_header_validity, &key, &value, ColBlockHeader);
-        }
-        for (key, value) in self.store.clone().iter(ColBlock) {
-            // Block Hash is valid
-            self.check(&block_hash_validity, &key, &value, ColBlock);
-            // Block Header for current Block exists
-            self.check(&block_header_exists, &key, &value, ColBlock);
-            // Block Height is greater or equal to tail, or to Genesis Height
-            self.check(&block_height_cmp_tail, &key, &value, ColBlock);
-        }
-        for (key, value) in self.store.clone().iter(ColChunks) {
-            // Chunk Hash is valid
-            self.check(&chunk_hash_validity, &key, &value, ColChunks);
-            // Block for current Chunk exists
-            self.check(&block_of_chunk_exists, &key, &value, ColChunks);
-            // There is a State Root in the Trie
-            self.check(&chunks_state_roots_in_trie, &key, &value, ColChunks);
-        }
-        for shard_id in 0..self.shard_tries.tries.len() {
-            println!("{}", shard_id);
-            // TODO ??
-        }
-    }
-
-    fn check(
-        &mut self,
-        f: &dyn Fn(&StoreValidator, &[u8], &[u8]) -> Result<(), String>,
-        key: &[u8],
-        value: &[u8],
-        col: DBCol,
-    ) {
-        let result = f(self, key, value);
-        self.tests += 1;
-        match result {
-            Ok(_) => {}
-            Err(msg) => self.errors.push(ErrorMessage::new(col, msg)),
-        }
-    }
 }
